requirements/common.txt
django==5.1
djangorestframework==3.15.2
django-allauth==0.63.6
dj-rest-auth==6.0.0
requests==2.32.3
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
drf-yasg==1.21.5